requirements.txt
logur==0.6.0
PySimpleGUI==4.60.3